+++ packages/bits/src/lib/search/search.component.ts
@@ -140,11 +140,8 @@ export class SearchComponent implements IFilterPub {
     public onCancel(): void {
         this.value.set("");
         this.cancel.emit(this.value());
-        this.captureFocus.set(false);
-        queueMicrotask(() => {
-            this.captureFocus.set(true);
-            this.focusChange.emit(true);
-        });
+        this.captureFocus.set(true);
+        this.focusChange.emit(true);
     }
 
     public onFocusChange(event: boolean): void {