+++ docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [21.0.4] 📅 2026-09-16
+
+### Fixes
+
+- `@nova-ui/bits` | A11y fixies for Home Summary
+- `@nova-ui/bits` | A11y fixies for Search component
+- `@nova-ui/bits` | new accessible blue for active menu button text
+
 ## [21.0.3] 📅 2026-08-04
 
 ### Fixes